+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 * New: Added new API method to get enabled and disabled favorites.
     * public Airing[] GetPotentialFavoriteAirings(Favorite Favorite);
 * Fix: Aliases without a non-alias would cause an NPE when searching.
-* Fix: Schedule Direct aliasing logic was applied backwards.
+* Fix: Schedules Direct aliasing logic was applied backwards.
 * Carny is now multi-threaded and highly optimized.
+* New: Schedules Direct movie length is now imported.
+* New: Schedules Direct alternative channel logos can now be used by changing the property sdepg_core/use_alternate_logos=false to true. 
 
 ## Version 9.0.14 (2017-03-18)
 * New: Added new API methods for in progress sports tracking using Schedules Direct.